pyproject.toml
[project]
name = "home"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "uy_sun", email = "hmy0119@gmail.com" }
]
dependencies = [
    "django>=5.0.6",
    "strawberry-graphql-django>=0.46.1",
    "django-mptt>=0.16.0",
    "pillow>=10.4.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/home"]
